Provide useful info on Eclipse API failures
The Eclipse API service doesn't seem to be remarkable stable as far as
returning meaninful data. For example, at the moment, despite Eclipse
2021-09 not yet being released, it is the value returned by the API;
however, the `packages` array is totally empty. Of course, this check
fails with that input. And it likely should; however, the current error
ends up being something about list indices and that's not fun. So this
at least ensures that somewhat useful data is provided.