fix(eclipse): update fetching is broken after recent update
This returns to the format of parsing the `release.xml` file and getting
the URLs from that. The downloads API now requires a `release_name`
which is not particularly helpful because this tool's whole purpose is
to find the name of the release.

reverts: ffed5c1d8e52edced1974f560815f64a8654c70e